[Toolchain] Break apart Webpack into smaller pieces
