+++ package.json
@@ -10,7 +10,9 @@
     "acceptance": "scripts/acceptance.sh",
     "acceptance-record": "scripts/acceptance_record.sh",
     "assets": "scripts/assets.sh",
-    "bundle-report": "NODE_ENV=production ANALYZE_BUNDLE=true webpack",
+    "build": "yarn webpack",
+    "bundle-report": "NODE_ENV=production ANALYZE_BUNDLE=true yarn build",
+    "bundle-report:dev": "NODE_ENV=development ANALYZE_BUNDLE=true yarn build",
     "mocha": "scripts/mocha.sh",
     "precommit": "lint-staged",
     "prepare": "patch-package",
@@ -26,7 +28,7 @@
     "sync-schema": "rm -rf data && mkdir data && graphql-fetch-schema https://metaphysics-staging.artsy.net -o data",
     "test": "scripts/test.sh",
     "type-check": "tsc --pretty --noEmit",
-    "webpack": "webpack"
+    "webpack": "node --max_old_space_size=4096 -r dotenv/config -r @babel/register node_modules/.bin/webpack --config ./webpack"
   },
   "prettier": {
     "bracketSpacing": true,
@@ -238,7 +240,7 @@
     "@types/react": "^16.4.13",
     "@types/react-dom": "^16.0.7",
     "@types/styled-system": "^3.0.4",
-    "@types/webpack": "^4.4.11",
+    "@types/webpack": "^4.4.24",
     "@types/webpack-env": "^1.13.6",
     "babel-core": "^7.0.0-bridge.0",
     "babel-jest": "^23.6.0",
@@ -286,15 +288,19 @@
     "sinon": "^1.17.1",
     "source-map-support": "^0.4.14",
     "static-extend": "^0.1.2",
+    "style-loader": "^0.23.1",
+    "stylus-loader": "^3.0.2",
     "supertest": "^2.0.1",
     "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.15.0",
     "tslint-config-standard": "^8.0.1",
     "tslint-react": "^3.6.0",
-    "webpack": "^4.20.2",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-middleware": "^3.4.0",
+    "uglifyjs-webpack-plugin": "^2.1.1",
+    "webpack": "^4.29.0",
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-middleware": "^3.5.1",
     "webpack-hot-middleware": "^2.24.3",
+    "webpack-merge": "^4.2.1",
     "webpack-notifier": "^1.7.0"
   },
   "license": "MIT",
