+++ scripts/assets.sh
@@ -5,7 +5,7 @@ set -ex
 rm -rf public
 mkdir public
 mkdir public/assets
-NODE_ENV=production node --max_old_space_size=4096 node_modules/.bin/webpack
+NODE_ENV=production yarn build
 stylus \
   $(find src/desktop/assets src/mobile/assets -name '*.styl') \
   --compress \
