+++ src/desktop/apps/order2/server.js
@@ -1,9 +1,7 @@
 import express from "express"
 import { checkoutFlow } from "desktop/apps/order2/routes"
 
-const app = (module.exports = express())
+export const app = express()
 
 app.get("/orders/:orderID*", checkoutFlow)
 app.get("/order2/:orderID*", checkoutFlow)
-
-export default app
