+++ src/index.js
@@ -35,11 +35,11 @@ cache.setup(() => {
   // again - this prevents a sustained broken state when gravity returns a
   // 502 during force startup.
   artsyXapp.on("error", err => {
-    error(`
+    console.error(`
 Could not start Force because it could not set up the xapp token, this is likely
-due to \`API_URL\`, \`CLIENT_ID\` and \`CLIENT_SECRET\` not being set, but 
+due to \`API_URL\`, \`CLIENT_ID\` and \`CLIENT_SECRET\` not being set, but
 also could be gravity being down.`)
-    error(err)
+    console.error(err)
     process.exit()
   })
   // Get an xapp token
