+++ webpack/envs/ci.js
@@ -1,0 +1,6 @@
+// @ts-check
+
+export const ciConfig = {
+  mode: "development",
+  devtool: false,
+}
